requirements.txt
conan>=2
PyYAML
